+++ tests/test_tweets.py
@@ -12,7 +12,7 @@ def test_update_status(twitter_credentials):
         status_id = update_status(twitter_credentials, status=status)
         return status_id
 
-    assert test_flow().result().result() == status
+    assert test_flow() == status
 
 
 def test_update_status_missing(twitter_credentials):
@@ -22,7 +22,7 @@ def test_update_status_missing(twitter_credentials):
         return status_id
 
     with pytest.raises(ValueError, match="One of text or media_ids"):
-        test_flow().result().result()
+        test_flow()
 
 
 def test_update_status_media(twitter_credentials):
@@ -33,7 +33,7 @@ def test_update_status_media(twitter_credentials):
         status_id = update_status(twitter_credentials, media_ids=media_ids)
         return status_id
 
-    assert test_flow().result().result() == media_ids[0]
+    assert test_flow() == media_ids[0]
 
 
 def test_get_status(twitter_credentials):
@@ -44,4 +44,4 @@ def test_get_status(twitter_credentials):
         status = get_status(status_id, twitter_credentials)
         return status
 
-    assert test_flow().result().result() == status_id
+    assert test_flow() == status_id