+++ CHANGELOG.md
@@ -10,6 +10,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 ### Changed
+- Updated tests to be compatible with core Prefect library (v2.0b9) and bumped required version - [#29](https://github.com/PrefectHQ/prefect-twitter/pull/29)
+- Converted `TwitterCredentials` into a `Block` - [#29](https://github.com/PrefectHQ/prefect-twitter/pull/29)
 
 ### Deprecated